move direction classes to recursive_backtracking namespace
