+++ lib/maze_magic.rb
@@ -5,10 +5,6 @@ require 'maze_magic/grid'
 require 'maze_magic/maze_generator'
 require 'maze_magic/maze_generator/randomnes'
 
-%w(north south east west).each do |file|
-  require "maze_magic/maze_generator/#{file}"
-end
-
 %w(horizontal_wall vertical_wall passage edge).each do |file|
   require "maze_magic/#{file}"
 end
