--- lib/maze_magic/maze_generator/east.rb
@@ -1,23 +1,0 @@
-module MazeMagic
-  module MazeGenerator
-    class East
-      include Singleton
-
-      def direction_x
-        1
-      end
-
-      def  direction_y
-        0
-      end
-
-      def oposite
-        West.instance
-      end
-
-      def to_i
-        4
-      end
-    end
-  end
-end
