+++ lib/maze_magic/maze_generator/recursive_backtracking/east.rb
@@ -1,0 +1,25 @@
+module MazeMagic
+  module MazeGenerator
+    class RecursiveBacktracking
+      class East
+        include Singleton
+
+        def direction_x
+          1
+        end
+
+        def  direction_y
+          0
+        end
+
+        def oposite
+          West.instance
+        end
+
+        def to_i
+          4
+        end
+      end
+    end
+  end
+end
