+++ lib/maze_magic/maze_generator/recursive_backtracking/west.rb
@@ -1,0 +1,25 @@
+module MazeMagic
+  module MazeGenerator
+    class RecursiveBacktracking
+      class West
+        include Singleton
+
+        def direction_x
+          -1
+        end
+
+        def  direction_y
+          0
+        end
+
+        def oposite
+          East.instance
+        end
+
+        def to_i
+          8
+        end
+      end
+    end
+  end
+end
