--- lib/maze_magic/maze_generator/south.rb
@@ -1,23 +1,0 @@
-module MazeMagic
-  module MazeGenerator
-    class South
-      include Singleton
-
-      def direction_x
-        0
-      end
-
-      def  direction_y
-        1
-      end
-
-      def oposite
-        North.instance
-      end
-
-      def to_i
-        2
-      end
-    end
-  end
-end
