--- lib/maze_magic/maze_generator/west.rb
@@ -1,23 +1,0 @@
-module MazeMagic
-  module MazeGenerator
-    class West
-      include Singleton
-
-      def direction_x
-        -1
-      end
-
-      def  direction_y
-        0
-      end
-
-      def oposite
-        East.instance
-      end
-
-      def to_i
-        8
-      end
-    end
-  end
-end
